支持通过INLINE环境变量区分处理是否需要将css、js文件内联至html中的情况

+++ template/build/tasks/remove-redundant-folder.js
@@ -1,9 +1,18 @@
+const gulp = require('gulp')
 const fse = require('fs-extra')
 
-const { join, projectName } = require('../utils')
+const {
+    join,
+    projectName,
+    shouldInline,
+} = require('../utils')
 
 module.exports = () => {
-    // 因为添加了inline属性的link/script标签所引用的js和css文件会被内联到html中，所以产物中不需要保留css和js目录
-    fse.remove(join(`/${projectName}/js`))
-    return fse.remove(join(`/${projectName}/css`))
+    if (shouldInline) {
+        // 因为添加了inline属性的link/script标签所引用的js和css文件会被内联到html中，所以产物中不需要保留css和js目录
+        fse.removeSync(join(`/${projectName}/js`))
+        return fse.remove(join(`/${projectName}/css`))
+    }
+    // 这里具体gulp.src的文件不重要，只是为了返回一个"空"的gulp任务
+    return gulp.src(join('/src/site.data.config.js'))
 }
